requirements.txt
Flask
flask-cors
gunicorn
numpy
Pillow
cv2
https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp36-cp36m-linux_x86_64.whl